Add network check while loading data so we will not crash, rename some variables

+++ app/src/main/java/com/ratushny/poeasisto/ninja/data/model/itemoverview/NinjaItemLines.kt
@@ -5,5 +5,5 @@ data class NinjaItemLines(
     val icon: String,
     val sparkline: NinjaItemReceiveSparkLine,
     val chaosValue: Float,
-    val exaltedValue: Double
+    val exaltedValue: Float
 )
